requirements.txt
altair==5.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
community==1.0.0b1
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
dhlab==2.26.1
executing==1.2.0
Flask==2.3.2
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-louvain==0.16
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
rdflib==7.0.0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.25.0
streamlit-agraph==0.0.45
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.0
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wordcloud==1.9.2
zipp==3.16.2
